[MFLP-46] feat: Functional login system accomplished
A JWT authentication system was implemented in Flask (PyJWT), considering decorators for route protection and JWT management. Also, a user and token management was implemented in React using Zustand. Route protection was implemented for security and permission purposes. Finally, the navbar now behaves depending if the user is a guest or a logged in user, so he can register or log in (guest) or else see his recipes (W.I.P.) or logout (auth user).